fix: failed to write to local storage
When creating a pending Safe the the tx’s value was a bigInt 0n, this was causing the json Serializer to crash. The JSON standard doesn’t know how to treat BigInts, but. thankfully the parse an serialize functions have a second argument that can be used for custom types.

Whenever we stumble on a BigInt we store it in the storage as an object of {__type: ‘bigint’, __value: ‘0x…’}
Later when we load the key, we transform it back to a BigInt(0x…)