fix: UI not updating if signing from walletconnect
If another safe is owner of a safe and we use walletconnect to sign transaction the UI was not updating. It turns out we need to port the  UncheckedJsonRpcSigner from ethers v5.

The UncheckedJsonRpcSigner fakes a transaction receipt without actually having one,
and this is needed for transactions that don’t
actually end on-chain.