fix: hydration error on dashboard
When in dev mode and on a chain different than eth or sepolia every refresh display a “Hydration failed error”. It turns out that the governance module is being rendered on the server, where in the client it return null.
This is caused by the use of useChainId hook which returns a default chain on the server, but on the client the actual chain the user is on.
With the isLayoutReady check we make sure that this component actually only renders when ran on the client.