feat: hide speed up button for non-signed txs
There are a lot of wallets out there that ignore the specified nonce. If we offer the option to speed up when the tx is not signed we can end up with multiple transactions in the wallet. The moment the user manages to speed the one tx that is stuck all subsequent txs will be executed and this could lead to potential loss of funds(funds sent multiple times). Because of this we no longer display the speed up option, but rather ask the user to manually speed up through the wallet.