fix: wrong orderClass used
context:
Trust the appData.metadata.orderClass.orderClass to match the form used (Swap->market, Limit->limit, Twap->twap).
More context:
The the order class in the API is what the order was considered during placement.
Before, there was a difference between SWAP and LIMIT.
Basically SWAP had a fee, LIMIT did not.
When we moved all orders to not have a pre-defined fee, everything became a LIMIT order from the protocol point of view.

.

.