refactor: switch to pull_request_target (#3966)
Github actions on forks fail, because they require secrets that are not
available. By relying on pull_request_target maintainers will be able to
start those action workflows and they shouldn’t fail.

From the github docs:
This event runs in the context of the base of the pull request, rather
than in the context of the merge commit, as the pull_request event does.
This prevents execution of unsafe code from the head of the pull request
that could alter your repository or steal any secrets you use in your
workflow. This event allows your workflow to do things like label or
comment on pull requests from forks.