fix: twap parse error
In situations where the parent twap order is generating more than 10
child orders CGW isn’t fetching the child orders. Because of that it
can’t calculate the executed buy and sell amounts and returns null for
it. The defined gateway types were not reflecting this and the returned
null value was causing the parsing to fail.

.