fix: hide twap status when we don’t know the state
TWAP orders are missing a backend service where we can ask for the
status. In reality to determine the current twap status we have to
iterrate through all twap parts and look at their individual statuses.
This is not a viable solution for twap orders with many parts,
so we’ve decided to not display the status until CoW comes up with a
backend service.