fix: useIsExpiredSwap causing too many rerenders
The Hook was using IntervalCounter and despite the state of the hook
not changing, the interval counter was being update, which in turn
caused every component using this hook to rerender.

Since this hook is used in TXSummary it meant that the whole TXhistory
or TX Queue were rerendered every 10s degrading the performance.