refactor: usage of getTransactionDetails
I’ve moved the usage of getTransactionDetails to rtk query. This has
the benefit of caching the result and multiple components relying on
that data don’t have to re-request the same data as it will be available
in the redux store.

In addition to this, this change fixes a bug in the txDetails component.
We were polling for the swap status, but due to the way
useIntervalCounter works, we were causing component rerenders no matter
if the tx was a swap or not.