fix: keep ofac query data for 24h
Currently once we do the ofac check the data will be discarded after 60s
(if no active listeners) and whenever we navigate to a page that needs
that data a new call to the rpc is going to be made. With this change
the data will be kept for up to 24h (if no active listener).