fix: useLoadChains should follow the next url
The getConfigs function was using the client gateway sdk to fetch the
chains. This call is limited to 20 items and we never try to fetch more.
The config service now has more than 20 networks and we are missing them
in the url. It turned out however that it is very difficult to provide
the {offset, limit} arguments to the getChainInfo on the sdk. Something
seems wrong and generated next url is wrong.

That’s why I opted out of using the sdk for this call and instead
directly fetch the results and follow the next prop on the returned
object until there is nothing more.