Add SafeNet Check Summary to Signature Form
This PR adds the SafeNet checks to the transaction execution flow in the
wallet interface.

Currently, the checks are shown for any Safe that has SafeNet enabled
(i.e. has the SafeNet guard set on a supported chain). The checks are
done using the SafeNet service backend.

Results are grouped into categories which allows the UI to show high
level summaries based on a collection of guarantees as well as grouping
any unknown guarantees into an "other" category.

Note that I implemented SafeTx hash computation as I couldn't find a
ulitity method that did it for me anywhere - which was surprising. If
there is an existing utility method for this - please let me know and I
will change my code to use it.